fix: remove and separate study list element
